requirements.txt
requests
retry
pyyaml
pillow